create an EventBus composition local for use within Compose
